Augment MethodPointer Relocation
If the method to be materialized by the TR_MethodPointer relocation
record is the same as the method at the associated inlined site index,
then there is no need to explicitly materialize it at relo time; the
method from the inlined site is sufficient.

This subtlety is important because the method pointer in BreakPoint
guards are relocated using TR_MethodPointer, and it needs to be the same
as the method in the inlining table in the J9JITExceptionTable.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>