Disable Recompilation under FSD & Sync Compilation
An app thread that blocks on compilation releases VM Access. If class
redefinition occurs during a recompilation, the application thread
can no longer OSR out to the interpreter; it is forced to return to
the oldStartPC (to jump to a helper) which may not necessarily be
valid. Therefore, this commit disables recompilation if the JIT is
doing synchronous compilation in FSD mode.

In a normal run, compilations occur asynchronously. However, there
are circumstances where a recompilation must be done synchronously.
They have to do with method invalidation, specifically preexistence
and HCR. However, in FSD, preexistence is disabled, and HCR is
implemented using involuntary OSR. Therefore, the only
circumstance under which sync recompilations can occur under FSD is
when explicitly requested (via counts or disableAsyncCompilation).

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>