Use Portable AOT target cpu in CRIU mode
If the JVM Is in CRIU mode and checkpointing is allowed, then the JIT
should be limited to the same processor features as those used in
Portable AOT mode. This is because, the restore run may not be on the
same machine as the one that created the snapshot; thus the JIT code
must be portable.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>