Support SVM AOT compilation of invokeHandle/invokeDynamic
Support unresolved dispatch of invokeHandle/invokeDynamic for SVM AOT
compilations. No additional relocation/validation records are needed as
the method type table entry / callsite entry addresses are not emitted
in the generated code. Additionally, the polymorphic signature of the
dummy linkToStatic resolved method is entirely determined by the callee
and thus the combination of the callee's SVM ID and Class Chain
validations are sufficient to validate the polymorphic signature.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>