Add External Options Metadata
This commit adds the ExternalOptionsMetadata struct as well as a table
of this struct to contain information about External Options that the
JIT processes. This commit also adds a method to find and possibly
consume the options, caching the arg index. Finally, this commit adds
some documentation to describe how to add a new external option.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>